roles: Notify when role is not locked
Summary: Useful to get a notification even if the role isn't locked so that we can alert if no instance is locking a role.
Up until now, we'd only trigger a callback with locked == false if we previously had success and called locked == true.
This change will trigger a callback if we request a role but don't end up locking it.