roles: Wait for a synchronous lock
Summary: We have issues during role exchange where the new role owner tries to lock the role before the old one can release it. This results in the new owner backing of for (by default) 10 seconds. This change will allow us to wait (by default) 1 second to watch the role mutex to see if it gets released, if it is then the new owner can lock it immediately.