roles: Unlock roles in parallel
Summary: When an instance has a lot of roles and then loses them on being re-ranked, it can take a while to unlock them all so that the new instance can lock them. With this MR we add two improvements, on a re-rank event we process all the unlocks in parallel and when the session is being closed we don't bother unlocking the roles as the session will expire them for us.